Expose testing related stubs with fixtures
In the gem, we are stubbing all external API request at the moment
But if any of our users wants to use those then they have to copy
paste this or something, so let's take this one to the next label
and expose all of our internal stubs as test helper.
