+++ spec/ribose/leaderboard_spec.rb
@@ -10,10 +10,4 @@ RSpec.describe Ribose::Leaderboard do
       expect(leaderboard.first.login).to eq("john.doe")
     end
   end
-
-  def stub_ribose_leaderboard_api
-    stub_api_response(
-      :get, "activity_point/leaderboard", filename: "leaderboard", status: 200
-    )
-  end
 end
