+++ spec/ribose/stream_spec.rb
@@ -11,8 +11,4 @@ RSpec.describe Ribose::Stream do
       expect(stream.notifications.first.data.first.info.action).to eq("create")
     end
   end
-
-  def stub_ribose_stream_list_api
-    stub_api_response(:get, "stream", filename: "stream", status: 200)
-  end
 end
