+++ spec/spec_helper.rb
@@ -1,15 +1,13 @@
 require "webmock/rspec"
 require "bundler/setup"
 require "ribose"
+require "ribose/rspec"
 
 Dir["./spec/support/**/*.rb"].sort.each { |file| require file }
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = ".rspec_status"
-
-  # Ribose API, Response stubs
-  config.include Ribose::FakeRiboseApi
 
   config.expect_with :rspec do |c|
     c.syntax = :expect
